Method get products by name. Changed use settings. Debug crud. Use ProblemDetail.

+++ src/Services/Catalog/Catalog.Api/Repositories/ProductRepository.cs
@@ -23,32 +23,41 @@ namespace Catalog.Api.Repositories
             return await _context.Products.Find(x => x.Id == id).FirstOrDefaultAsync();
         }
 
-        public async Task<IEnumerable<Product>> GetProductByCategoryAsync(string categoryName)
+        public async Task<IEnumerable<Product>> GetProductByNameAsync(string name)
         {
-            FilterDefinition<Product> filter = Builders<Product>.Filter.ElemMatch(p => p.Category, categoryName);
+            FilterDefinition<Product> filter = Builders<Product>.Filter.Eq(p => p.Name, name);
             return await _context.Products.Find(filter).ToListAsync();
         }
 
-        public async Task<IEnumerable<Product>> GetProductByName(string name)
+        public async Task<IEnumerable<Product>> GetProductByCategoryAsync(string categoryName)
         {
-            FilterDefinition<Product> filter = Builders<Product>.Filter.ElemMatch(p => p.Name, name);
+            FilterDefinition<Product> filter = Builders<Product>.Filter.Eq(p => p.Category, categoryName);
             return await _context.Products.Find(filter).ToListAsync();
         }
 
         public async Task CreateProductAsync(Product product)
         {
+            product.Id = string.Empty;
             await _context.Products.InsertOneAsync(product);
         }
 
         public async Task<bool> UpdateProductAsync(Product product)
         {
+            var existProduct = await _context.Products.Find(x => x.Id == product.Id).FirstOrDefaultAsync();
+            if (existProduct is null)
+                throw new KeyNotFoundException($"The product whith identifier '{product.Id}' is not found.");
+
             var updateResult = await _context.Products.ReplaceOneAsync(x => x.Id == product.Id, product);
             return updateResult.IsAcknowledged && updateResult.ModifiedCount > 0;
         }
 
         public async Task<bool> DeleteProductAsync(string id)
         {
-            FilterDefinition<Product> filter = Builders<Product>.Filter.ElemMatch(p => p.Id, id);
+            FilterDefinition<Product> filter = Builders<Product>.Filter.Eq(p => p.Id, id);
+            var existProduct = await _context.Products.Find(filter).FirstOrDefaultAsync();
+            if (existProduct is null)
+                throw new KeyNotFoundException($"The product whith identifier '{id}' is not found.");
+
             DeleteResult deleteResult = await _context.Products.DeleteOneAsync(filter);
             return deleteResult.IsAcknowledged && deleteResult.DeletedCount > 0;
         }
